requirements.txt
Flask==3.0.0
requests==2.28.2
openai==1.23.2
google-cloud-tasks==2.15.0
google-cloud-datastore==2.19.0
google-cloud-ndb==2.3.1
google-api-core==2.17.0

Authlib===1.3.1
pyOpenSSL==24.1.0
Flask-Login==0.6.3